+++ internal/controlplane/server_test.go
@@ -19,14 +19,19 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/base64"
+	"encoding/json"
+	"errors"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 	"time"
 
+	"github.com/cenkalti/backoff/v4"
+	"github.com/google/go-github/v63/github"
 	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
@@ -78,10 +83,11 @@ func init() {
 func newDefaultServer(
 	t *testing.T,
 	mockStore *mockdb.MockStore,
+	// TODO: can be removed?
 	mockRepoSvc *mock_reposvc.MockRepositoryService,
 	mockPropSvc *mock_service.MockPropertiesService,
 	ghClientFactory ghclient.GitHubClientFactory,
-) (*Server, events.Interface) {
+) *Server {
 	t.Helper()
 
 	evt, err := events.Setup(context.Background(), &serverconfig.EventConfig{
@@ -135,7 +141,7 @@ func newDefaultServer(
 		cryptoEngine:  eng,
 	}
 
-	return server, evt
+	return server
 }
 
 func generateTokenKey(t *testing.T) string {
@@ -166,4 +172,67 @@ func TestWebhook(t *testing.T) {
 		t.Fatalf("Failed to get webhook: %v", err)
 	}
 	defer resp.Body.Close()
+}
+
+func TestHandleRequestTooLarge(t *testing.T) {
+	t.Parallel()
+
+	handler := func(w http.ResponseWriter, r *http.Request) {
+		// Reading the body to trigger the limit check
+		_, err := io.ReadAll(r.Body)
+		if err != nil {
+			var maxErr *http.MaxBytesError
+			if errors.As(err, &maxErr) {
+				// this is the expected error
+				http.Error(w, "request body too large", http.StatusRequestEntityTooLarge)
+				return
+			}
+			// this is an unexpected error
+			http.Error(w, "internal server error", http.StatusInternalServerError)
+			return
+		}
+
+		// a 200 isn't necessary, but it would fail the test if the handler didn't return 413
+		w.WriteHeader(http.StatusOK)
+	}
+
+	ts := httptest.NewServer(withMaxSizeMiddleware(http.HandlerFunc(handler)))
+
+	event := github.PackageEvent{
+		Action: github.String("published"),
+		Repo: &github.Repository{
+			ID:   github.Int64(12345),
+			Name: github.String("stacklok/minder"),
+		},
+		Org: &github.Organization{
+			Login: github.String("stacklok"),
+		},
+	}
+	packageJson, err := json.Marshal(event)
+	require.NoError(t, err, "failed to marshal package event")
+
+	maliciousBody := strings.NewReader(strings.Repeat("1337", 100000000))
+	maliciousBodyReader := io.MultiReader(maliciousBody, maliciousBody, maliciousBody, maliciousBody, maliciousBody)
+	_ = packageJson
+
+	req, err := http.NewRequest("POST", ts.URL, maliciousBodyReader)
+	require.NoError(t, err, "failed to create request")
+
+	req.Header.Add("X-GitHub-Event", "meta")
+	req.Header.Add("X-GitHub-Delivery", "12345")
+	req.Header.Add("Content-Type", "application/json")
+	resp, err := httpDoWithRetry(ts.Client(), req)
+	require.NoError(t, err, "failed to make request")
+	require.Equal(t, http.StatusRequestEntityTooLarge, resp.StatusCode, "unexpected status code")
+}
+
+func httpDoWithRetry(client *http.Client, req *http.Request) (*http.Response, error) {
+	var resp *http.Response
+	err := backoff.Retry(func() error {
+		var err error
+		resp, err = client.Do(req)
+		return err
+	}, backoff.WithMaxRetries(backoff.NewConstantBackOff(time.Second), 3))
+
+	return resp, err
 }
