+++ internal/providers/github/webhook/errors.go
@@ -1,0 +1,41 @@
+// Copyright 2024 Stacklok, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package webhook
+
+import (
+	"errors"
+	"fmt"
+)
+
+// errRepoNotFound is returned when a repository is not found
+var errRepoNotFound = errors.New("repository not found")
+
+// errArtifactNotFound is returned when an artifact is not found
+var errArtifactNotFound = errors.New("artifact not found")
+
+// errArtifactVersionSkipped is returned when an artifact is skipped because it has no tags
+var errArtifactVersionSkipped = errors.New("artifact version skipped, has no tags")
+
+// errRepoIsPrivate is returned when a repository is private
+var errRepoIsPrivate = errors.New("repository is private")
+
+// errNotHandled is returned when a webhook event is not handled
+var errNotHandled = errors.New("webhook event not handled")
+
+// newErrNotHandled returns a new errNotHandled error
+func newErrNotHandled(smft string, args ...any) error {
+	msg := fmt.Sprintf(smft, args...)
+	return fmt.Errorf("%w: %s", errNotHandled, msg)
+}
