+++ internal/providers/github/webhook/test-payloads/branch-protection-configuration-disabled.json
