+++ internal/providers/github/webhook/test-payloads/installation-deleted.json
